+++ R/theme_gtsummary.R
@@ -22,7 +22,7 @@
 #'       - `tbl_summary()` Doesn't show percent symbol; use em-dash to separate IQR
 #'   - `"qjecon"` _The Quarterly Journal of Economics_ ___Under Development___
 #'       - `tbl_summary()` all percentages rounded to one decimal place
-#'       - `tbl_regression()`/`tbl_uvregression()` add significance stars with `add_significance_stars()`; hides CI and p-value from output and adds column for SE
+#'       - `tbl_regression()`/`tbl_uvregression()` add significance stars with `add_significance_stars()`; hides CI and p-value from output
 #' - `theme_gtsummary_compact()`
 #'   - tables printed with gt, flextable, kableExtra, or huxtable will be compact with smaller font size and reduced cell padding
 #' - `theme_gtsummary_printer(print_engine=)`
@@ -32,6 +32,9 @@
 #' - `theme_gtsummary_mean_sd()`
 #'   - Set default summary statistics to mean and standard deviation in `tbl_summary()`
 #'   - Set default continuous tests in `add_p()` to t-test and ANOVA
+#' - `theme_gtsummary_eda()`
+#'   - Set all continuous variables to summary type `"continuous2"` by default
+#'   - In `tbl_summary()` show the median, mean, IQR, SD, and Range by default
 #'
 #' Use `reset_gtsummary_theme()` to restore the default settings
 #'
@@ -65,7 +68,8 @@ NULL
 #' @export
 #' @param journal String indicating the journal theme to follow. One of
 #' `c("jama", "lancet", "nejm", "qjecon")`. Details below.
-theme_gtsummary_journal <- function(journal = c("jama", "lancet", "nejm", "qjecon"), set_theme = TRUE) {
+theme_gtsummary_journal <- function(journal = c("jama", "lancet", "nejm", "qjecon"),
+                                    set_theme = TRUE) {
   journal <- match.arg(journal)
   if (journal == "jama") {
     lst_theme <-
@@ -191,18 +195,27 @@ theme_gtsummary_journal <- function(journal = c("jama", "lancet", "nejm", "qjeco
           new_header_text <-
             paste(
               x$table_styling$header %>% filter(.data$column == "estimate") %>% pull(.data$label),
-              "**(SE)**"
+              "**(SE)**",
+              sep = " "
             )
+
+          estimate_footnote <-
+            x$table_styling$footnote_abbrev %>%
+            filter(.data$column %in% "estimate") %>%
+            filter(dplyr::row_number() == dplyr::n(), !is.na(.data$footnote)) %>%
+            dplyr::pull(.data$footnote) %>%
+            c("SE = Standard Error") %>%
+            paste(collapse = ", ")
 
           x %>%
             add_significance_stars(
-              pattern = "{estimate}{stars} ({std.error})",
+              pattern = "{estimate}{stars}\n({std.error})",
               hide_se = TRUE
             ) %>%
             # update column header
             modify_header(list(estimate = new_header_text)) %>%
-            # add CI abbreviation footnote
-            modify_footnote(estimate ~ "SE = Standard Error", abbreviation = TRUE)
+            # add SE abbreviation footnote
+            modify_footnote(estimate ~ estimate_footnote, abbreviation = TRUE)
         }
       )
   }
@@ -353,6 +366,23 @@ theme_gtsummary_mean_sd <- function(set_theme = TRUE) {
     "add_p.tbl_summary-attr:test.continuous_by2" = "t.test",
     "add_p.tbl_summary-attr:test.continuous" = "aov",
     "add_p.tbl_svysummary-attr:test.continuous" = "svy.t.test"
+  )
+
+  if (set_theme == TRUE) set_gtsummary_theme(lst_theme)
+  return(invisible(lst_theme))
+}
+
+# ------------------------------------------------------------------------------
+#' @rdname theme_gtsummary
+#' @export
+theme_gtsummary_eda <- function(set_theme = TRUE) {
+
+  lst_theme <- list(
+    "pkgwide-str:theme_name" = "Exploratory Data Analysis",
+    "tbl_summary-str:default_con_type" = "continuous2",
+    "tbl_summary-str:continuous_stat" =
+      c("{median} ({p25}, {p75})", "{mean} ({sd})", "{min}, {max}"),
+    "tbl_summary-fn:percent_fun" = function(x) style_percent(x, digits = 1)
   )
 
   if (set_theme == TRUE) set_gtsummary_theme(lst_theme)